Avoid casting to raw types in some scenarios
Istead of recursively capturing a return type of method, apply custom transformation
replacing non-denotable types with their denotable parents (we do not use intersections
as return type of serializer-strategies, so typevars are effectively the only such types).

For bonus readability, try to capture requested type and see, if wildcards weren't needed
(this have to be done after serializer-strategy is is called, because those don't have
complete picture).